requirements.txt
streamlit
python-dotenv
faiss-cpu
langchain-google-genai
langchain-community
PyPDF2
langchain-core
python-docx
markdown